Break SSS limits. A small mod in the image_operations.py blurring algorithm makes the blur really intense without affecting quality.     Minor changes: Fixed a bug in image viewer's save as button. TODO: add filters in the dialog, for saving default png, etc. Also added a refresh button. It surprisingly reveals a defect. You have to change tabs for the refresh to take effect. Added two methods in the ImageView class to handle mouse input...  that don't really seem to work, for some reason. Probably a bug. But it's like the code's never executed. Will see.

+++ makehuman/plugins/4_rendering_povray/mh2povray.py
@@ -803,12 +803,12 @@ def povrayProcessSSS(stuffs, outDir, settings, progressCallback = None):
         os.path.join(outDir, '%s_sss_bluelmap.png' % stuffs[0].name))
     # green channel
     progress(0.4*nextpb)
-    lmap2 = imgop.blurred(lmap, int(10*sssa),13, lambda p: progress((0.4+0.3*p)*nextpb))
+    lmap2 = imgop.blurred(lmap, 10*sssa, 13, lambda p: progress((0.4+0.3*p)*nextpb))
     imgop.compose([black,lmap2,black]).save(
         os.path.join(outDir, '%s_sss_greenlmap.png' % stuffs[0].name))
     # red channel
     progress(0.7*nextpb)
-    lmap2 = imgop.blurred(lmap2, int(20*sssa),13, lambda p: progress((0.7+0.3*p)*nextpb))
+    lmap2 = imgop.blurred(lmap2, 20*sssa, 13, lambda p: progress((0.7+0.3*p)*nextpb))
     imgop.compose([lmap2,black,black]).save(
         os.path.join(outDir, '%s_sss_redlmap.png' % stuffs[0].name))
     progbase = nextpb
@@ -817,11 +817,11 @@ def povrayProcessSSS(stuffs, outDir, settings, progressCallback = None):
         # Export blurred bump maps
         lmap = imgop.Image(os.path.join(stuffs[0].bump[0], stuffs[0].bump[1]))
         lmap = imgop.getChannel(lmap,1)
-        lmap = imgop.blurred(lmap, int(float(lmap.width/1024)*5*sssa), 15,
+        lmap = imgop.blurred(lmap, float(lmap.width)/1024*5*sssa, 15,
                              lambda p: progress(progbase+0.5*p*(1-progbase)))
         progress(progbase+0.5*(1-progbase))
         lmap.save(os.path.join(outDir, '%s_sss_greenbump.png' % stuffs[0].name))
-        lmap = imgop.blurred(lmap, int(float(lmap.width/1024)*10*sssa), 15,
+        lmap = imgop.blurred(lmap, float(lmap.width)/1024*10*sssa, 15,
                              lambda p: progress(progbase+(0.5+0.5*p)*(1-progbase)))
         lmap.save(os.path.join(outDir, '%s_sss_redbump.png' % stuffs[0].name))
         progress(1.0)
